% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddDimensionGroupResponse}
\alias{AddDimensionGroupResponse}
\title{AddDimensionGroupResponse Object}
\usage{
AddDimensionGroupResponse(dimensionGroups = NULL)
}
\arguments{
\item{dimensionGroups}{All groups of a dimension after adding a group to that dimension}
}
\value{
AddDimensionGroupResponse object
}
\description{
AddDimensionGroupResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of adding a group.
}
