% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AppendCellsRequest}
\alias{AppendCellsRequest}
\title{AppendCellsRequest Object}
\usage{
AppendCellsRequest(rows = NULL, fields = NULL, sheetId = NULL)
}
\arguments{
\item{rows}{The data to append}

\item{fields}{The fields of CellData that should be updated}

\item{sheetId}{The sheet ID to append the data to}
}
\value{
AppendCellsRequest object
}
\description{
AppendCellsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Adds new cells after the last row with data in a sheet,inserting new rows into the sheet if necessary.
}
