% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AppendDimensionRequest}
\alias{AppendDimensionRequest}
\title{AppendDimensionRequest Object}
\usage{
AppendDimensionRequest(length = NULL, sheetId = NULL,
  dimension = NULL)
}
\arguments{
\item{length}{The number of rows or columns to append}

\item{sheetId}{The sheet to append rows or columns to}

\item{dimension}{Whether rows or columns should be appended}
}
\value{
AppendDimensionRequest object
}
\description{
AppendDimensionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Appends rows or columns to the end of a sheet.
}
\concept{AppendDimensionRequest functions}
