% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BandedRange}
\alias{BandedRange}
\title{BandedRange Object}
\usage{
BandedRange(range = NULL, bandedRangeId = NULL, rowProperties = NULL,
  columnProperties = NULL)
}
\arguments{
\item{range}{The range over which these properties are applied}

\item{bandedRangeId}{The id of the banded range}

\item{rowProperties}{Properties for row bands}

\item{columnProperties}{Properties for column bands}
}
\value{
BandedRange object
}
\description{
BandedRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A banded (alternating colors) range in a sheet.
}
