% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BasicChartSpec}
\alias{BasicChartSpec}
\title{BasicChartSpec Object}
\usage{
BasicChartSpec(threeDimensional = NULL, axis = NULL, chartType = NULL,
  interpolateNulls = NULL, series = NULL, legendPosition = NULL,
  compareMode = NULL, domains = NULL, lineSmoothing = NULL,
  headerCount = NULL, stackedType = NULL)
}
\arguments{
\item{threeDimensional}{True to make the chart 3D}

\item{axis}{The axis on the chart}

\item{chartType}{The type of the chart}

\item{interpolateNulls}{If some values in a series are missing, gaps may appear in the chart (e}

\item{series}{The data this chart is visualizing}

\item{legendPosition}{The position of the chart legend}

\item{compareMode}{The behavior of tooltips and data highlighting when hovering on data and}

\item{domains}{The domain of data this is charting}

\item{lineSmoothing}{Gets whether all lines should be rendered smooth or straight by default}

\item{headerCount}{The number of rows or columns in the data that are 'headers'}

\item{stackedType}{The stacked type for charts that support vertical stacking}
}
\value{
BasicChartSpec object
}
\description{
BasicChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The specification for a basic chart.  See BasicChartType for the listof charts this supports.
}
