% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BatchClearValuesByDataFilterRequest}
\alias{BatchClearValuesByDataFilterRequest}
\title{BatchClearValuesByDataFilterRequest Object}
\usage{
BatchClearValuesByDataFilterRequest(dataFilters = NULL)
}
\arguments{
\item{dataFilters}{The DataFilters used to determine which ranges to clear}
}
\value{
BatchClearValuesByDataFilterRequest object
}
\description{
BatchClearValuesByDataFilterRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for clearing more than one range selected by aDataFilter in a spreadsheet.
}
\seealso{
Other BatchClearValuesByDataFilterRequest functions: \code{\link{spreadsheets.values.batchClearByDataFilter}}
}
