% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BatchClearValuesByDataFilterResponse}
\alias{BatchClearValuesByDataFilterResponse}
\title{BatchClearValuesByDataFilterResponse Object}
\usage{
BatchClearValuesByDataFilterResponse(clearedRanges = NULL,
  spreadsheetId = NULL)
}
\arguments{
\item{clearedRanges}{The ranges that were cleared, in A1 notation}

\item{spreadsheetId}{The spreadsheet the updates were applied to}
}
\value{
BatchClearValuesByDataFilterResponse object
}
\description{
BatchClearValuesByDataFilterResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response when clearing a range of values selected withDataFilters in a spreadsheet.
}
