% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BatchGetValuesByDataFilterRequest}
\alias{BatchGetValuesByDataFilterRequest}
\title{BatchGetValuesByDataFilterRequest Object}
\usage{
BatchGetValuesByDataFilterRequest(valueRenderOption = NULL,
  dateTimeRenderOption = NULL, majorDimension = NULL, dataFilters = NULL)
}
\arguments{
\item{valueRenderOption}{How values should be represented in the output}

\item{dateTimeRenderOption}{How dates, times, and durations should be represented in the output}

\item{majorDimension}{The major dimension that results should use}

\item{dataFilters}{The data filters used to match the ranges of values to retrieve}
}
\value{
BatchGetValuesByDataFilterRequest object
}
\description{
BatchGetValuesByDataFilterRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for retrieving a range of values in a spreadsheet selected by aset of DataFilters.
}
\seealso{
Other BatchGetValuesByDataFilterRequest functions: \code{\link{spreadsheets.values.batchGetByDataFilter}}
}
