% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BatchUpdateValuesByDataFilterResponse}
\alias{BatchUpdateValuesByDataFilterResponse}
\title{BatchUpdateValuesByDataFilterResponse Object}
\usage{
BatchUpdateValuesByDataFilterResponse(totalUpdatedColumns = NULL,
  spreadsheetId = NULL, totalUpdatedRows = NULL, responses = NULL,
  totalUpdatedSheets = NULL, totalUpdatedCells = NULL)
}
\arguments{
\item{totalUpdatedColumns}{The total number of columns where at least one cell in the column was}

\item{spreadsheetId}{The spreadsheet the updates were applied to}

\item{totalUpdatedRows}{The total number of rows where at least one cell in the row was updated}

\item{responses}{The response for each range updated}

\item{totalUpdatedSheets}{The total number of sheets where at least one cell in the sheet was}

\item{totalUpdatedCells}{The total number of cells updated}
}
\value{
BatchUpdateValuesByDataFilterResponse object
}
\description{
BatchUpdateValuesByDataFilterResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response when updating a range of values in a spreadsheet.
}
