% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{Border}
\alias{Border}
\title{Border Object}
\usage{
Border(color = NULL, width = NULL, style = NULL)
}
\arguments{
\item{color}{The color of the border}

\item{width}{The width of the border, in pixels}

\item{style}{The style of the border}
}
\value{
Border object
}
\description{
Border Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A border along a cell.
}
