% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CandlestickChartSpec}
\alias{CandlestickChartSpec}
\title{CandlestickChartSpec Object}
\usage{
CandlestickChartSpec(domain = NULL, data = NULL)
}
\arguments{
\item{domain}{The domain data (horizontal axis) for the candlestick chart}

\item{data}{The Candlestick chart data}
}
\value{
CandlestickChartSpec object
}
\description{
CandlestickChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A <a href='/chart/interactive/docs/gallery/candlestickchart'>candlestick chart</a>.
}
