% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CandlestickData}
\alias{CandlestickData}
\title{CandlestickData Object}
\usage{
CandlestickData(highSeries = NULL, lowSeries = NULL, closeSeries = NULL,
  openSeries = NULL)
}
\arguments{
\item{highSeries}{The range data (vertical axis) for the high/maximum value for each}

\item{lowSeries}{The range data (vertical axis) for the low/minimum value for each candle}

\item{closeSeries}{The range data (vertical axis) for the close/final value for each candle}

\item{openSeries}{The range data (vertical axis) for the open/initial value for each}
}
\value{
CandlestickData object
}
\description{
CandlestickData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Candlestick chart data, each containing the low, open, close, and highvalues for a series.
}
