% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CandlestickDomain}
\alias{CandlestickDomain}
\title{CandlestickDomain Object}
\usage{
CandlestickDomain(data = NULL, reversed = NULL)
}
\arguments{
\item{data}{The data of the CandlestickDomain}

\item{reversed}{True to reverse the order of the domain values (horizontal axis)}
}
\value{
CandlestickDomain object
}
\description{
CandlestickDomain Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The domain of a CandlestickChart.
}
