% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CandlestickSeries}
\alias{CandlestickSeries}
\title{CandlestickSeries Object}
\usage{
CandlestickSeries(data = NULL)
}
\arguments{
\item{data}{The data of the CandlestickSeries}
}
\value{
CandlestickSeries object
}
\description{
CandlestickSeries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The series of a CandlestickData.
}
