% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CellData}
\alias{CellData}
\title{CellData Object}
\usage{
CellData(pivotTable = NULL, userEnteredFormat = NULL, note = NULL,
  effectiveFormat = NULL, dataValidation = NULL,
  userEnteredValue = NULL, effectiveValue = NULL,
  formattedValue = NULL, textFormatRuns = NULL, hyperlink = NULL)
}
\arguments{
\item{pivotTable}{A pivot table anchored at this cell}

\item{userEnteredFormat}{The format the user entered for the cell}

\item{note}{Any note on the cell}

\item{effectiveFormat}{The effective format being used by the cell}

\item{dataValidation}{A data validation rule on the cell, if any}

\item{userEnteredValue}{The value the user entered in the cell}

\item{effectiveValue}{The effective value of the cell}

\item{formattedValue}{The formatted value of the cell}

\item{textFormatRuns}{Runs of rich text applied to subsections of the cell}

\item{hyperlink}{A hyperlink this cell points to, if any}
}
\value{
CellData object
}
\description{
CellData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data about a specific cell.
}
\concept{CellData functions}
