% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ChartSpec}
\alias{ChartSpec}
\title{ChartSpec Object}
\usage{
ChartSpec(title = NULL, altText = NULL, titleTextPosition = NULL,
  histogramChart = NULL, candlestickChart = NULL, bubbleChart = NULL,
  waterfallChart = NULL, fontName = NULL, maximized = NULL,
  treemapChart = NULL, hiddenDimensionStrategy = NULL,
  subtitleTextFormat = NULL, subtitle = NULL,
  subtitleTextPosition = NULL, backgroundColor = NULL,
  basicChart = NULL, orgChart = NULL, pieChart = NULL,
  titleTextFormat = NULL)
}
\arguments{
\item{title}{The title of the chart}

\item{altText}{The alternative text that describes the chart}

\item{titleTextPosition}{The title text position}

\item{histogramChart}{A histogram chart specification}

\item{candlestickChart}{A candlestick chart specification}

\item{bubbleChart}{A bubble chart specification}

\item{waterfallChart}{A waterfall chart specification}

\item{fontName}{The name of the font to use by default for all chart text (e}

\item{maximized}{True to make a chart fill the entire space in which it's rendered with}

\item{treemapChart}{A treemap chart specification}

\item{hiddenDimensionStrategy}{Determines how the charts will use hidden rows or columns}

\item{subtitleTextFormat}{The subtitle text format}

\item{subtitle}{The subtitle of the chart}

\item{subtitleTextPosition}{The subtitle text position}

\item{backgroundColor}{The background color of the entire chart}

\item{basicChart}{A basic chart specification, can be one of many kinds of charts}

\item{orgChart}{An org chart specification}

\item{pieChart}{A pie chart specification}

\item{titleTextFormat}{The title text format}
}
\value{
ChartSpec object
}
\description{
ChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The specifications of a chart.
}
\concept{ChartSpec functions}
