% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CreateDeveloperMetadataRequest}
\alias{CreateDeveloperMetadataRequest}
\title{CreateDeveloperMetadataRequest Object}
\usage{
CreateDeveloperMetadataRequest(developerMetadata = NULL)
}
\arguments{
\item{developerMetadata}{The developer metadata to create}
}
\value{
CreateDeveloperMetadataRequest object
}
\description{
CreateDeveloperMetadataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to create developer metadata.
}
