% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CreateDeveloperMetadataResponse}
\alias{CreateDeveloperMetadataResponse}
\title{CreateDeveloperMetadataResponse Object}
\usage{
CreateDeveloperMetadataResponse(developerMetadata = NULL)
}
\arguments{
\item{developerMetadata}{The developer metadata that was created}
}
\value{
CreateDeveloperMetadataResponse object
}
\description{
CreateDeveloperMetadataResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response from creating developer metadata.
}
