% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CutPasteRequest}
\alias{CutPasteRequest}
\title{Google Sheets API Objects 
Reads and writes Google Sheets.}
\usage{
CutPasteRequest(source = NULL, pasteType = NULL, destination = NULL)
}
\arguments{
\item{source}{The source data to cut}

\item{pasteType}{What kind of data to paste}

\item{destination}{The top-left coordinate where the data should be pasted}
}
\value{
CutPasteRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2018-08-28 13:36:11
filename: ~/Desktop/delete_me//googlesheetsv4.auto/R/sheets_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
CutPasteRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Moves data from the source to the destination.
}
\concept{CutPasteRequest functions}
