% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DataFilter}
\alias{DataFilter}
\title{DataFilter Object}
\usage{
DataFilter(developerMetadataLookup = NULL, a1Range = NULL,
  gridRange = NULL)
}
\arguments{
\item{developerMetadataLookup}{Selects data associated with the developer metadata matching the criteria}

\item{a1Range}{Selects data that matches the specified A1 range}

\item{gridRange}{Selects data that matches the range described by the GridRange}
}
\value{
DataFilter object
}
\description{
DataFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Filter that describes what data should be selected or returned from arequest.
}
\concept{DataFilter functions}
