% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DataFilterValueRange}
\alias{DataFilterValueRange}
\title{DataFilterValueRange Object}
\usage{
DataFilterValueRange(dataFilter = NULL, majorDimension = NULL,
  values = NULL)
}
\arguments{
\item{dataFilter}{The data filter describing the location of the values in the spreadsheet}

\item{majorDimension}{The major dimension of the values}

\item{values}{The data to be written}
}
\value{
DataFilterValueRange object
}
\description{
DataFilterValueRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A range of values whose location is specified by a DataFilter.
}
