% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DataValidationRule}
\alias{DataValidationRule}
\title{DataValidationRule Object}
\usage{
DataValidationRule(strict = NULL, inputMessage = NULL, condition = NULL,
  showCustomUi = NULL)
}
\arguments{
\item{strict}{True if invalid data should be rejected}

\item{inputMessage}{A message to show the user when adding data to the cell}

\item{condition}{The condition that data in the cell must match}

\item{showCustomUi}{True if the UI should be customized based on the kind of condition}
}
\value{
DataValidationRule object
}
\description{
DataValidationRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A data validation rule.
}
