% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DateTimeRule}
\alias{DateTimeRule}
\title{DateTimeRule Object}
\usage{
DateTimeRule(type = NULL)
}
\arguments{
\item{type}{The type of date-time grouping to apply}
}
\value{
DateTimeRule object
}
\description{
DateTimeRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Allows you to organize the date-time values in a source data column intobuckets based on selected parts of their date or time values. For example,consider a pivot table showing sales transactions by date:    +----------+--------------+    | Date     | SUM of Sales |    +----------+--------------+    | 1/1/2017 |      $621.14 |    | 2/3/2017 |      $708.84 |    | 5/8/2017 |      $326.84 |    ...    +----------+--------------+Applying a date-time group rule with a DateTimeRuleType of YEAR_MONTHresults in the following pivot table.    +--------------+--------------+    | Grouped Date | SUM of Sales |    +--------------+--------------+    | 2017-Jan     |   $53,731.78 |    | 2017-Feb     |   $83,475.32 |    | 2017-Mar     |   $94,385.05 |    ...    +--------------+--------------+
}
