% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteDeveloperMetadataRequest}
\alias{DeleteDeveloperMetadataRequest}
\title{DeleteDeveloperMetadataRequest Object}
\usage{
DeleteDeveloperMetadataRequest(dataFilter = NULL)
}
\arguments{
\item{dataFilter}{The data filter describing the criteria used to select which developer}
}
\value{
DeleteDeveloperMetadataRequest object
}
\description{
DeleteDeveloperMetadataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to delete developer metadata.
}
