% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteDeveloperMetadataResponse}
\alias{DeleteDeveloperMetadataResponse}
\title{DeleteDeveloperMetadataResponse Object}
\usage{
DeleteDeveloperMetadataResponse(deletedDeveloperMetadata = NULL)
}
\arguments{
\item{deletedDeveloperMetadata}{The metadata that was deleted}
}
\value{
DeleteDeveloperMetadataResponse object
}
\description{
DeleteDeveloperMetadataResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response from deleting developer metadata.
}
