% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteDimensionGroupRequest}
\alias{DeleteDimensionGroupRequest}
\title{DeleteDimensionGroupRequest Object}
\usage{
DeleteDimensionGroupRequest(range = NULL)
}
\arguments{
\item{range}{The range of the group to be deleted}
}
\value{
DeleteDimensionGroupRequest object
}
\description{
DeleteDimensionGroupRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes a group over the specified range by decrementing the depth of thedimensions in the range.For example, assume the sheet has a depth-1 group over B:E and a depth-2group over C:D. Deleting a group over D:E leaves the sheet with adepth-1 group over B:D and a depth-2 group over C:C.
}
