% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteDimensionGroupResponse}
\alias{DeleteDimensionGroupResponse}
\title{DeleteDimensionGroupResponse Object}
\usage{
DeleteDimensionGroupResponse(dimensionGroups = NULL)
}
\arguments{
\item{dimensionGroups}{All groups of a dimension after deleting a group from that dimension}
}
\value{
DeleteDimensionGroupResponse object
}
\description{
DeleteDimensionGroupResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of deleting a group.
}
