% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeveloperMetadata}
\alias{DeveloperMetadata}
\title{DeveloperMetadata Object}
\usage{
DeveloperMetadata(metadataValue = NULL, metadataKey = NULL,
  metadataId = NULL, location = NULL, visibility = NULL)
}
\arguments{
\item{metadataValue}{Data associated with the metadata's key}

\item{metadataKey}{The metadata key}

\item{metadataId}{The spreadsheet-scoped unique ID that identifies the metadata}

\item{location}{The location where the metadata is associated}

\item{visibility}{The metadata visibility}
}
\value{
DeveloperMetadata object
}
\description{
DeveloperMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Developer metadata associated with a location or object in a spreadsheet.Developer metadata may be used to associate arbitrary data with variousparts of a spreadsheet and will remain associated at those locations as theymove around and the spreadsheet is edited.  For example, if developermetadata is associated with row 5 and another row is then subsequentlyinserted above row 5, that original metadata will still be associated withthe row it was first associated with (what is now row 6). If the associatedobject is deleted its metadata is deleted too.
}
