% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeveloperMetadataLocation}
\alias{DeveloperMetadataLocation}
\title{DeveloperMetadataLocation Object}
\usage{
DeveloperMetadataLocation(dimensionRange = NULL, spreadsheet = NULL,
  sheetId = NULL, locationType = NULL)
}
\arguments{
\item{dimensionRange}{Represents the row or column when metadata is associated with}

\item{spreadsheet}{True when metadata is associated with an entire spreadsheet}

\item{sheetId}{The ID of the sheet when metadata is associated with an entire sheet}

\item{locationType}{The type of location this object represents}
}
\value{
DeveloperMetadataLocation object
}
\description{
DeveloperMetadataLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A location where metadata may be associated in a spreadsheet.
}
