% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DimensionGroup}
\alias{DimensionGroup}
\title{DimensionGroup Object}
\usage{
DimensionGroup(collapsed = NULL, range = NULL, depth = NULL)
}
\arguments{
\item{collapsed}{This field is true if this group is collapsed}

\item{range}{The range over which this group exists}

\item{depth}{The depth of the group, representing how many groups have a range that}
}
\value{
DimensionGroup object
}
\description{
DimensionGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group over an interval of rows or columns on a sheet, which can contain orbe contained within other groups. A group can be collapsed or expanded as aunit on the sheet.
}
