% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{EmbeddedChart}
\alias{EmbeddedChart}
\title{EmbeddedChart Object}
\usage{
EmbeddedChart(spec = NULL, chartId = NULL, position = NULL)
}
\arguments{
\item{spec}{The specification of the chart}

\item{chartId}{The ID of the chart}

\item{position}{The position of the chart}
}
\value{
EmbeddedChart object
}
\description{
EmbeddedChart Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A chart embedded in a sheet.
}
