% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_objects.R
\name{File.capabilities}
\alias{File.capabilities}
\title{File.capabilities Object}
\usage{
File.capabilities(canAddChildren = NULL,
  canChangeCopyRequiresWriterPermission = NULL,
  canChangeViewersCanCopyContent = NULL, canComment = NULL,
  canCopy = NULL, canDelete = NULL, canDownload = NULL, canEdit = NULL,
  canListChildren = NULL, canMoveItemIntoTeamDrive = NULL,
  canMoveTeamDriveItem = NULL, canReadRevisions = NULL,
  canReadTeamDrive = NULL, canRemoveChildren = NULL, canRename = NULL,
  canShare = NULL, canTrash = NULL, canUntrash = NULL)
}
\arguments{
\item{canAddChildren}{Whether the current user can add children to this folder}

\item{canChangeCopyRequiresWriterPermission}{Whether the current user can change the copyRequiresWriterPermission restriction of this file}

\item{canChangeViewersCanCopyContent}{Deprecated}

\item{canComment}{Whether the current user can comment on this file}

\item{canCopy}{Whether the current user can copy this file}

\item{canDelete}{Whether the current user can delete this file}

\item{canDownload}{Whether the current user can download this file}

\item{canEdit}{Whether the current user can edit this file}

\item{canListChildren}{Whether the current user can list the children of this folder}

\item{canMoveItemIntoTeamDrive}{Whether the current user can move this item into a Team Drive}

\item{canMoveTeamDriveItem}{Whether the current user can move this Team Drive item by changing its parent}

\item{canReadRevisions}{Whether the current user can read the revisions resource of this file}

\item{canReadTeamDrive}{Whether the current user can read the Team Drive to which this file belongs}

\item{canRemoveChildren}{Whether the current user can remove children from this folder}

\item{canRename}{Whether the current user can rename this file}

\item{canShare}{Whether the current user can modify the sharing settings for this file}

\item{canTrash}{Whether the current user can move this file to trash}

\item{canUntrash}{Whether the current user can restore this file from trash}
}
\value{
File.capabilities object
}
\description{
File.capabilities Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Capabilities the current user has on this file. Each capability corresponds to a fine-grained action that a user may take.
}
\seealso{
Other File functions: \code{\link{File.appProperties}},
  \code{\link{File.properties}}, \code{\link{File}},
  \code{\link{files.copy}}
}
