% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{FindReplaceRequest}
\alias{FindReplaceRequest}
\title{FindReplaceRequest Object}
\usage{
FindReplaceRequest(sheetId = NULL, matchCase = NULL, allSheets = NULL,
  includeFormulas = NULL, matchEntireCell = NULL, find = NULL,
  searchByRegex = NULL, replacement = NULL, range = NULL)
}
\arguments{
\item{sheetId}{The sheet to find/replace over}

\item{matchCase}{True if the search is case sensitive}

\item{allSheets}{True to find/replace over all sheets}

\item{includeFormulas}{True if the search should include cells with formulas}

\item{matchEntireCell}{True if the find value should match the entire cell}

\item{find}{The value to search}

\item{searchByRegex}{True if the find value is a regex}

\item{replacement}{The value to use as the replacement}

\item{range}{The range to find/replace over}
}
\value{
FindReplaceRequest object
}
\description{
FindReplaceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Finds and replaces data in cells over a range, sheet, or all sheets.
}
