% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{GetSpreadsheetByDataFilterRequest}
\alias{GetSpreadsheetByDataFilterRequest}
\title{GetSpreadsheetByDataFilterRequest Object}
\usage{
GetSpreadsheetByDataFilterRequest(dataFilters = NULL,
  includeGridData = NULL)
}
\arguments{
\item{dataFilters}{The DataFilters used to select which ranges to retrieve from}

\item{includeGridData}{True if grid data should be returned}
}
\value{
GetSpreadsheetByDataFilterRequest object
}
\description{
GetSpreadsheetByDataFilterRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for retrieving a Spreadsheet.
}
\seealso{
Other GetSpreadsheetByDataFilterRequest functions: \code{\link{spreadsheets.getByDataFilter}}
}
