% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{gsheet_clear_ss_values}
\alias{gsheet_clear_ss_values}
\title{Execute Google Sheets `spreadsheets.values.clear method`}
\usage{
gsheet_clear_ss_values(spreadsheet_id, range, simplify_vector = FALSE)
}
\arguments{
\item{spreadsheet_id}{The ID of the spreadsheet to retrieve data from.}

\item{range}{The A1 notation of the values to clear}

\item{simplify_vector}{passed to jsonlite::fromJSON, defaults to FALSE}
}
\value{
list with spreadsheetId and clearedRange
}
\description{
[v4 API Doc](https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/get)
}
