% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{gsheet_update_ss_values}
\alias{gsheet_update_ss_values}
\title{Execute Google Sheets Method `spreadsheets.values.update`}
\usage{
gsheet_update_ss_values(spreadsheet_id, value_range,
  value_input_option = "USER_ENTERED",
  include_values_in_response = FALSE,
  response_value_render_option = "FORMATTED_VALUE",
  response_date_time_render_option = "SERIAL_NUMBER",
  simplify_vector = FALSE)
}
\arguments{
\item{spreadsheet_id}{atomic character vector length 1; The ID of the
spreadsheet to update.}

\item{value_range}{ValueRange object; to be passed to value request body}

\item{value_input_option}{atomic character length 1; How the input data
should be interpreted.}

\item{include_values_in_response}{atomic logical length 1; Determines if the
update response should include the values of the cells that were updated.
By default, responses do not include the updated values. If the range to
write was larger than than the range actually written, the response will
include all values in the requested range (excluding trailing empty rows
and columns).}

\item{response_value_render_option}{atomic character length 1; Determines
how values in the response should be rendered. The default render option
is ValueRenderOption.FORMATTED_VALUE.}

\item{response_date_time_render_option}{atomic character length 1; Determines
how dates, times, and durations in the response should be rendered. This
is ignored if responseValueRenderOption is FORMATTED_VALUE. The default
dateTime render option is [DateTimeRenderOption.SERIAL_NUMBER].}

\item{simplify_vector}{passed to jsonlite::fromJSON, defaults to FALSE}
}
\value{
update_values_response object (list)
}
\description{
[v4 API Doc](https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/update)
}
