% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{HistogramChartSpec}
\alias{HistogramChartSpec}
\title{HistogramChartSpec Object}
\usage{
HistogramChartSpec(legendPosition = NULL, bucketSize = NULL,
  outlierPercentile = NULL, showItemDividers = NULL, series = NULL)
}
\arguments{
\item{legendPosition}{The position of the chart legend}

\item{bucketSize}{By default the bucket size (the range of values stacked in a single}

\item{outlierPercentile}{The outlier percentile is used to ensure that outliers do not adversely}

\item{showItemDividers}{Whether horizontal divider lines should be displayed between items in each}

\item{series}{The series for a histogram may be either a single series of values to be}
}
\value{
HistogramChartSpec object
}
\description{
HistogramChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A <a href='/chart/interactive/docs/gallery/histogram'>histogram chart</a>.A histogram chart groups data items into bins, displaying each bin as acolumn of stacked items.  Histograms are used to display the distributionof a dataset.  Each column of items represents a range into which thoseitems fall.  The number of bins can be chosen automatically or specifiedexplicitly.
}
