% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{HistogramRule}
\alias{HistogramRule}
\title{HistogramRule Object}
\usage{
HistogramRule(interval = NULL, start = NULL, end = NULL)
}
\arguments{
\item{interval}{The size of the buckets that are created}

\item{start}{The minimum value at which items are placed into buckets}

\item{end}{The maximum value at which items are placed into buckets}
}
\value{
HistogramRule object
}
\description{
HistogramRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Allows you to organize the numeric values in a source data column intobuckets of a constant size. All values from HistogramRule.start toHistogramRule.end are placed into groups of sizeHistogramRule.interval. In addition, all values belowHistogramRule.start are placed in one group, and all values aboveHistogramRule.end are placed in another. OnlyHistogramRule.interval is required, though if HistogramRule.startand HistogramRule.end are both provided, HistogramRule.start mustbe less than HistogramRule.end. For example, a pivot table showingaverage purchase amount by age that has 50+ rows:    +-----+-------------------+    | Age | AVERAGE of Amount |    +-----+-------------------+    | 16  |            $27.13 |    | 17  |             $5.24 |    | 18  |            $20.15 |    ...    +-----+-------------------+could be turned into a pivot table that looks like the one below byapplying a histogram group rule with a HistogramRule.start of 25,an HistogramRule.interval of 20, and an HistogramRule.endof 65.    +-------------+-------------------+    | Grouped Age | AVERAGE of Amount |    +-------------+-------------------+    | < 25        |            $19.34 |    | 25-45       |            $31.43 |    | 45-65       |            $35.87 |    | > 65        |            $27.55 |    +-------------+-------------------+    | Grand Total |            $29.12 |    +-------------+-------------------+
}
