% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{HistogramSeries}
\alias{HistogramSeries}
\title{HistogramSeries Object}
\usage{
HistogramSeries(barColor = NULL, data = NULL)
}
\arguments{
\item{barColor}{The color of the column representing this series in each bucket}

\item{data}{The data for this histogram series}
}
\value{
HistogramSeries object
}
\description{
HistogramSeries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A histogram series containing the series color and data.
}
