% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{LineStyle}
\alias{LineStyle}
\title{LineStyle Object}
\usage{
LineStyle(type = NULL, width = NULL)
}
\arguments{
\item{type}{The dash type of the line}

\item{width}{The thickness of the line, in px}
}
\value{
LineStyle object
}
\description{
LineStyle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties that describe the style of a line.
}
