% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ManualRule}
\alias{ManualRule}
\title{ManualRule Object}
\usage{
ManualRule(groups = NULL)
}
\arguments{
\item{groups}{The list of group names and the corresponding items from the source data}
}
\value{
ManualRule object
}
\description{
ManualRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Allows you to manually organize the values in a source data column intobuckets with names of your choosing. For example, a pivot table thataggregates population by state:    +-------+-------------------+    | State | SUM of Population |    +-------+-------------------+    | AK    |               0.7 |    | AL    |               4.8 |    | AR    |               2.9 |    ...    +-------+-------------------+could be turned into a pivot table that aggregates population by time zoneby providing a list of groups (for example, groupName = 'Central',items = ['AL', 'AR', 'IA', ...]) to a manual group rule.Note that a similar effect could be achieved by adding a time zone columnto the source data and adjusting the pivot table.    +-----------+-------------------+    | Time Zone | SUM of Population |    +-----------+-------------------+    | Central   |             106.3 |    | Eastern   |             151.9 |    | Mountain  |              17.4 |    ...    +-----------+-------------------+
}
