% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ManualRuleGroup}
\alias{ManualRuleGroup}
\title{ManualRuleGroup Object}
\usage{
ManualRuleGroup(groupName = NULL, items = NULL)
}
\arguments{
\item{groupName}{The group name, which must be a string}

\item{items}{The items in the source data that should be placed into this group}
}
\value{
ManualRuleGroup object
}
\description{
ManualRuleGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group name and a list of items from the source data that should be placedin the group with this name.
}
