% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{MatchedDeveloperMetadata}
\alias{MatchedDeveloperMetadata}
\title{MatchedDeveloperMetadata Object}
\usage{
MatchedDeveloperMetadata(developerMetadata = NULL, dataFilters = NULL)
}
\arguments{
\item{developerMetadata}{The developer metadata matching the specified filters}

\item{dataFilters}{All filters matching the returned developer metadata}
}
\value{
MatchedDeveloperMetadata object
}
\description{
MatchedDeveloperMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A developer metadata entry and the data filters specified in the originalrequest that matched it.
}
