% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{MatchedValueRange}
\alias{MatchedValueRange}
\title{MatchedValueRange Object}
\usage{
MatchedValueRange(valueRange = NULL, dataFilters = NULL)
}
\arguments{
\item{valueRange}{The values matched by the DataFilter}

\item{dataFilters}{The DataFilters from the request that matched the range of}
}
\value{
MatchedValueRange object
}
\description{
MatchedValueRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A value range that was matched by one or more data filers.
}
