% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_message}
\alias{my_message}
\title{Custom message log level}
\usage{
my_message(..., level = 2)
}
\arguments{
\item{...}{The message(s)}

\item{level}{The severity}
}
\description{
Custom message log level
}
\details{
0 = everything, 1 = debug, 2=normal, 3=important
}
\keyword{internal}
