% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{OrgChartSpec}
\alias{OrgChartSpec}
\title{OrgChartSpec Object}
\usage{
OrgChartSpec(tooltips = NULL, selectedNodeColor = NULL,
  parentLabels = NULL, nodeSize = NULL, labels = NULL, nodeColor = NULL)
}
\arguments{
\item{tooltips}{The data containing the tooltip for the corresponding node}

\item{selectedNodeColor}{The color of the selected org chart nodes}

\item{parentLabels}{The data containing the label of the parent for the corresponding node}

\item{nodeSize}{The size of the org chart nodes}

\item{labels}{The data containing the labels for all the nodes in the chart}

\item{nodeColor}{The color of the org chart nodes}
}
\value{
OrgChartSpec object
}
\description{
OrgChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An <a href='/chart/interactive/docs/gallery/orgchart'>org chart</a>.Org charts require a unique set of labels in labels and mayoptionally include parent_labels and tooltips.parent_labels contain, for each node, the label identifying the parentnode.  tooltips contain, for each node, an optional tooltip.For example, to describe an OrgChart with Alice as the CEO, Bob as thePresident (reporting to Alice) and Cathy as VP of Sales (also reporting toAlice), have labels contain 'Alice', 'Bob', 'Cathy',parent_labels contain '', 'Alice', 'Alice' and tooltips contain'CEO', 'President', 'VP Sales'.
}
