% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PivotGroup}
\alias{PivotGroup}
\title{PivotGroup Object}
\usage{
PivotGroup(sourceColumnOffset = NULL, sortOrder = NULL,
  valueBucket = NULL, showTotals = NULL, valueMetadata = NULL,
  groupRule = NULL, label = NULL, repeatHeadings = NULL)
}
\arguments{
\item{sourceColumnOffset}{The column offset of the source range that this grouping is based on}

\item{sortOrder}{The order the values in this group should be sorted}

\item{valueBucket}{The bucket of the opposite pivot group to sort by}

\item{showTotals}{True if the pivot table should include the totals for this grouping}

\item{valueMetadata}{Metadata about values in the grouping}

\item{groupRule}{The group rule to apply to this row/column group}

\item{label}{The labels to use for the row/column groups which can be customized}

\item{repeatHeadings}{True if the headings in this pivot group should be repeated}
}
\value{
PivotGroup object
}
\description{
PivotGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single grouping (either row or column) in a pivot table.
}
