% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PivotGroupRule}
\alias{PivotGroupRule}
\title{PivotGroupRule Object}
\usage{
PivotGroupRule(histogramRule = NULL, dateTimeRule = NULL,
  manualRule = NULL)
}
\arguments{
\item{histogramRule}{A HistogramRule}

\item{dateTimeRule}{A DateTimeRule}

\item{manualRule}{A ManualRule}
}
\value{
PivotGroupRule object
}
\description{
PivotGroupRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An optional setting on a PivotGroup that defines buckets for the valuesin the source data column rather than breaking out each individual value.Only one PivotGroup with a group rule may be added for each column inthe source data, though on any given column you may add both aPivotGroup that has a rule and a PivotGroup that does not.
}
