% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{RandomizeRangeRequest}
\alias{RandomizeRangeRequest}
\title{RandomizeRangeRequest Object}
\usage{
RandomizeRangeRequest(range = NULL)
}
\arguments{
\item{range}{The range to randomize}
}
\value{
RandomizeRangeRequest object
}
\description{
RandomizeRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Randomizes the order of the rows in a range.
}
