% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{Response}
\alias{Response}
\title{Response Object}
\usage{
Response(addFilterView = NULL, addBanding = NULL,
  addProtectedRange = NULL, duplicateSheet = NULL,
  updateEmbeddedObjectPosition = NULL,
  deleteConditionalFormatRule = NULL, deleteDimensionGroup = NULL,
  duplicateFilterView = NULL, addDimensionGroup = NULL,
  addChart = NULL, updateDeveloperMetadata = NULL,
  findReplace = NULL, addSheet = NULL,
  updateConditionalFormatRule = NULL, createDeveloperMetadata = NULL,
  addNamedRange = NULL, deleteDeveloperMetadata = NULL)
}
\arguments{
\item{addFilterView}{A reply from adding a filter view}

\item{addBanding}{A reply from adding a banded range}

\item{addProtectedRange}{A reply from adding a protected range}

\item{duplicateSheet}{A reply from duplicating a sheet}

\item{updateEmbeddedObjectPosition}{A reply from updating an embedded object's position}

\item{deleteConditionalFormatRule}{A reply from deleting a conditional format rule}

\item{deleteDimensionGroup}{A reply from deleting a dimension group}

\item{duplicateFilterView}{A reply from duplicating a filter view}

\item{addDimensionGroup}{A reply from adding a dimension group}

\item{addChart}{A reply from adding a chart}

\item{updateDeveloperMetadata}{A reply from updating a developer metadata entry}

\item{findReplace}{A reply from doing a find/replace}

\item{addSheet}{A reply from adding a sheet}

\item{updateConditionalFormatRule}{A reply from updating a conditional format rule}

\item{createDeveloperMetadata}{A reply from creating a developer metadata entry}

\item{addNamedRange}{A reply from adding a named range}

\item{deleteDeveloperMetadata}{A reply from deleting a developer metadata entry}
}
\value{
Response object
}
\description{
Response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single response from an update.
}
\concept{Response functions}
