% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{SearchDeveloperMetadataRequest}
\alias{SearchDeveloperMetadataRequest}
\title{SearchDeveloperMetadataRequest Object}
\usage{
SearchDeveloperMetadataRequest(dataFilters = NULL)
}
\arguments{
\item{dataFilters}{The data filters describing the criteria used to determine which}
}
\value{
SearchDeveloperMetadataRequest object
}
\description{
SearchDeveloperMetadataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to retrieve all developer metadata matching the set of specifiedcriteria.
}
\seealso{
Other SearchDeveloperMetadataRequest functions: \code{\link{spreadsheets.developerMetadata.search}}
}
