% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{SearchDeveloperMetadataResponse}
\alias{SearchDeveloperMetadataResponse}
\title{SearchDeveloperMetadataResponse Object}
\usage{
SearchDeveloperMetadataResponse(matchedDeveloperMetadata = NULL)
}
\arguments{
\item{matchedDeveloperMetadata}{The metadata matching the criteria of the search request}
}
\value{
SearchDeveloperMetadataResponse object
}
\description{
SearchDeveloperMetadataResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reply to a developer metadata search request.
}
