% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{Sheet}
\alias{Sheet}
\title{Sheet Object}
\usage{
Sheet(properties = NULL, developerMetadata = NULL, protectedRanges = NULL,
  conditionalFormats = NULL, columnGroups = NULL, basicFilter = NULL,
  merges = NULL, bandedRanges = NULL, charts = NULL, filterViews = NULL,
  rowGroups = NULL, data = NULL)
}
\arguments{
\item{properties}{The properties of the sheet}

\item{developerMetadata}{The developer metadata associated with a sheet}

\item{protectedRanges}{The protected ranges in this sheet}

\item{conditionalFormats}{The conditional format rules in this sheet}

\item{columnGroups}{All column groups on this sheet, ordered by increasing range start index,}

\item{basicFilter}{The filter on this sheet, if any}

\item{merges}{The ranges that are merged together}

\item{bandedRanges}{The banded (alternating colors) ranges on this sheet}

\item{charts}{The specifications of every chart on this sheet}

\item{filterViews}{The filter views in this sheet}

\item{rowGroups}{All row groups on this sheet, ordered by increasing range start index, then}

\item{data}{Data in the grid, if this is a grid sheet}
}
\value{
Sheet object
}
\description{
Sheet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A sheet in a spreadsheet.
}
