% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{SheetProperties}
\alias{SheetProperties}
\title{SheetProperties Object}
\usage{
SheetProperties(tabColor = NULL, index = NULL, sheetId = NULL,
  rightToLeft = NULL, hidden = NULL, gridProperties = NULL,
  sheetType = NULL, title = NULL)
}
\arguments{
\item{tabColor}{The color of the tab in the UI}

\item{index}{The index of the sheet within the spreadsheet}

\item{sheetId}{The ID of the sheet}

\item{rightToLeft}{True if the sheet is an RTL sheet instead of an LTR sheet}

\item{hidden}{True if the sheet is hidden in the UI, false if it's visible}

\item{gridProperties}{Additional properties of the sheet if this sheet is a grid}

\item{sheetType}{The type of sheet}

\item{title}{The name of the sheet}
}
\value{
SheetProperties object
}
\description{
SheetProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of a sheet.
}
