% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{SortSpec}
\alias{SortSpec}
\title{SortSpec Object}
\usage{
SortSpec(dimensionIndex = NULL, sortOrder = NULL)
}
\arguments{
\item{dimensionIndex}{The dimension the sort should be applied to}

\item{sortOrder}{The order data should be sorted}
}
\value{
SortSpec object
}
\description{
SortSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A sort order associated with a specific column or row.
}
