% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadsheets.developerMetadata.get}
\alias{spreadsheets.developerMetadata.get}
\title{Returns the developer metadata with the specified ID.The caller must specify the spreadsheet ID and the developer metadata'sunique metadataId.}
\usage{
spreadsheets.developerMetadata.get(spreadsheetId, metadataId)
}
\arguments{
\item{spreadsheetId}{The ID of the spreadsheet to retrieve metadata from}

\item{metadataId}{The ID of the developer metadata to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/spreadsheets
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/spreadsheets)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}
}
