% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadsheets.developerMetadata.search}
\alias{spreadsheets.developerMetadata.search}
\title{Returns all developer metadata matching the specified DataFilter.If the provided DataFilter represents a DeveloperMetadataLookup object,this will return all DeveloperMetadata entries selected by it. If theDataFilter represents a location in a spreadsheet, this will return alldeveloper metadata associated with locations intersecting that region.}
\usage{
spreadsheets.developerMetadata.search(SearchDeveloperMetadataRequest,
  spreadsheetId)
}
\arguments{
\item{SearchDeveloperMetadataRequest}{The \link{SearchDeveloperMetadataRequest} object to pass to this method}

\item{spreadsheetId}{The ID of the spreadsheet to retrieve metadata from}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/spreadsheets
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/spreadsheets)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}

Other SearchDeveloperMetadataRequest functions: \code{\link{SearchDeveloperMetadataRequest}}
}
