% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadsheets.getByDataFilter}
\alias{spreadsheets.getByDataFilter}
\title{Returns the spreadsheet at the given ID.The caller must specify the spreadsheet ID.This method differs from GetSpreadsheet in that it allows selectingwhich subsets of spreadsheet data to return by specifying adataFilters parameter.Multiple DataFilters can be specified.  Specifying one ormore data filters will return the portions of the spreadsheet thatintersect ranges matched by any of the filters.By default, data within grids will not be returned.You can include grid data one of two ways:* Specify a field mask listing your desired fields using the `fields` URLparameter in HTTP* Set the includeGridDataparameter to true.  If a field mask is set, the `includeGridData`parameter is ignoredFor large spreadsheets, it is recommended to retrieve only the specificfields of the spreadsheet that you want.}
\usage{
spreadsheets.getByDataFilter(GetSpreadsheetByDataFilterRequest, spreadsheetId)
}
\arguments{
\item{GetSpreadsheetByDataFilterRequest}{The \link{GetSpreadsheetByDataFilterRequest} object to pass to this method}

\item{spreadsheetId}{The spreadsheet to request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/spreadsheets
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/spreadsheets)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}

Other GetSpreadsheetByDataFilterRequest functions: \code{\link{GetSpreadsheetByDataFilterRequest}}
}
