% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadsheets.sheets.copyTo}
\alias{spreadsheets.sheets.copyTo}
\title{Copies a single sheet from a spreadsheet to another spreadsheet.Returns the properties of the newly created sheet.}
\usage{
spreadsheets.sheets.copyTo(CopySheetToAnotherSpreadsheetRequest, spreadsheetId,
  sheetId)
}
\arguments{
\item{CopySheetToAnotherSpreadsheetRequest}{The \link{CopySheetToAnotherSpreadsheetRequest} object to pass to this method}

\item{spreadsheetId}{The ID of the spreadsheet containing the sheet to copy}

\item{sheetId}{The ID of the sheet to copy}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/spreadsheets
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/spreadsheets)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}

Other CopySheetToAnotherSpreadsheetRequest functions: \code{\link{CopySheetToAnotherSpreadsheetRequest}}
}
