% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadsheets.values.append}
\alias{spreadsheets.values.append}
\title{Appends values to a spreadsheet. The input range is used to search forexisting data and find a 'table' within that range. Values will beappended to the next row of the table, starting with the first column ofthe table. See the[guide](/sheets/api/guides/values#appending_values)and[sample code](/sheets/api/samples/writing#append_values)for specific details of how tables are detected and data is appended.The caller must specify the spreadsheet ID, range, anda valueInputOption.  The `valueInputOption` onlycontrols how the input data will be added to the sheet (column-wise orrow-wise), it does not influence what cell the data starts being writtento.}
\usage{
spreadsheets.values.append(ValueRange, spreadsheetId, range,
  includeValuesInResponse = NULL, responseValueRenderOption = NULL,
  insertDataOption = NULL, valueInputOption = NULL,
  responseDateTimeRenderOption = NULL)
}
\arguments{
\item{ValueRange}{The \link{ValueRange} object to pass to this method}

\item{spreadsheetId}{The ID of the spreadsheet to update}

\item{range}{The A1 notation of a range to search for a logical table of data}

\item{includeValuesInResponse}{Determines if the update response should include the values}

\item{responseValueRenderOption}{Determines how values in the response should be rendered}

\item{insertDataOption}{How the input data should be inserted}

\item{valueInputOption}{How the input data should be interpreted}

\item{responseDateTimeRenderOption}{Determines how dates, times, and durations in the response should be}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/spreadsheets
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/spreadsheets)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}

Other ValueRange functions: \code{\link{ValueRange}},
  \code{\link{spreadsheets.values.update}}
}
