% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadsheets.values.batchClearByDataFilter}
\alias{spreadsheets.values.batchClearByDataFilter}
\title{Clears one or more ranges of values from a spreadsheet.The caller must specify the spreadsheet ID and one or moreDataFilters. Ranges matching any of the specified datafilters will be cleared.  Only values are cleared -- all other propertiesof the cell (such as formatting, data validation, etc..) are kept.}
\usage{
spreadsheets.values.batchClearByDataFilter(BatchClearValuesByDataFilterRequest,
  spreadsheetId)
}
\arguments{
\item{BatchClearValuesByDataFilterRequest}{The \link{BatchClearValuesByDataFilterRequest} object to pass to this method}

\item{spreadsheetId}{The ID of the spreadsheet to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/spreadsheets
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/spreadsheets)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}

Other BatchClearValuesByDataFilterRequest functions: \code{\link{BatchClearValuesByDataFilterRequest}}
}
