% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadsheets.values.batchGetByDataFilter}
\alias{spreadsheets.values.batchGetByDataFilter}
\title{Returns one or more ranges of values that match the specified data filters.The caller must specify the spreadsheet ID and one or moreDataFilters.  Ranges that match any of the data filters inthe request will be returned.}
\usage{
spreadsheets.values.batchGetByDataFilter(BatchGetValuesByDataFilterRequest,
  spreadsheetId)
}
\arguments{
\item{BatchGetValuesByDataFilterRequest}{The \link{BatchGetValuesByDataFilterRequest} object to pass to this method}

\item{spreadsheetId}{The ID of the spreadsheet to retrieve data from}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/spreadsheets
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/spreadsheets)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}

Other BatchGetValuesByDataFilterRequest functions: \code{\link{BatchGetValuesByDataFilterRequest}}
}
