% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadsheets.values.batchUpdateByDataFilter}
\alias{spreadsheets.values.batchUpdateByDataFilter}
\title{Sets values in one or more ranges of a spreadsheet.The caller must specify the spreadsheet ID,a valueInputOption, and one or moreDataFilterValueRanges.}
\usage{

  spreadsheets.values.batchUpdateByDataFilter(BatchUpdateValuesByDataFilterRequest,
  spreadsheetId)
}
\arguments{
\item{BatchUpdateValuesByDataFilterRequest}{The \link{BatchUpdateValuesByDataFilterRequest} object to pass to this method}

\item{spreadsheetId}{The ID of the spreadsheet to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/spreadsheets
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/spreadsheets)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}

Other BatchUpdateValuesByDataFilterRequest functions: \code{\link{BatchUpdateValuesByDataFilterRequest}}
}
