% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadsheets.values.clear}
\alias{spreadsheets.values.clear}
\title{Clears values from a spreadsheet.The caller must specify the spreadsheet ID and range.Only values are cleared -- all other properties of the cell (such asformatting, data validation, etc..) are kept.}
\usage{
spreadsheets.values.clear(ClearValuesRequest, spreadsheetId, range)
}
\arguments{
\item{ClearValuesRequest}{The \link{ClearValuesRequest} object to pass to this method}

\item{spreadsheetId}{The ID of the spreadsheet to update}

\item{range}{The A1 notation of the values to clear}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/spreadsheets
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/spreadsheets)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}

Other ClearValuesRequest functions: \code{\link{ClearValuesRequest}}
}
