% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadsheets.values.get}
\alias{spreadsheets.values.get}
\title{Returns a range of values from a spreadsheet.The caller must specify the spreadsheet ID and a range.}
\usage{
spreadsheets.values.get(spreadsheetId, range, valueRenderOption = NULL,
  dateTimeRenderOption = NULL, majorDimension = NULL)
}
\arguments{
\item{spreadsheetId}{The ID of the spreadsheet to retrieve data from}

\item{range}{The A1 notation of the values to retrieve}

\item{valueRenderOption}{How values should be represented in the output}

\item{dateTimeRenderOption}{How dates, times, and durations should be represented in the output}

\item{majorDimension}{The major dimension that results should use}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.readonly
\item https://www.googleapis.com/auth/spreadsheets
\item https://www.googleapis.com/auth/spreadsheets.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.file, https://www.googleapis.com/auth/drive.readonly, https://www.googleapis.com/auth/spreadsheets, https://www.googleapis.com/auth/spreadsheets.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}
}
