% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{spreadsheets_copy_file}
\alias{spreadsheets_copy_file}
\title{Copy a spreadsheet.}
\usage{
spreadsheets_copy_file(des_name, src_spreadsheetId)
}
\arguments{
\item{des_name}{The name of the new spreadsheet}

\item{src_spreadsheetId}{The ID of the original spreadsheet to copy}
}
\description{
Copy a spreadsheet.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}
}
