% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{spreadsheets_update}
\alias{spreadsheets_update}
\title{Sets values in a range of a spreadsheet.The caller must specify the spreadsheet ID, range.}
\usage{
spreadsheets_update(x, spreadsheetId, range)
}
\arguments{
\item{x}{The value to set in the spreadsheet}

\item{spreadsheetId}{The ID of the spreadsheet to update}

\item{range}{The A1 notation of the values to update}
}
\description{
Sets values in a range of a spreadsheet.The caller must specify the spreadsheet ID, range.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}
}
