% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{TextPosition}
\alias{TextPosition}
\title{TextPosition Object}
\usage{
TextPosition(horizontalAlignment = NULL)
}
\arguments{
\item{horizontalAlignment}{Horizontal alignment setting for the piece of text}
}
\value{
TextPosition object
}
\description{
TextPosition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Position settings for text.
}
