% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{TreemapChartColorScale}
\alias{TreemapChartColorScale}
\title{TreemapChartColorScale Object}
\usage{
TreemapChartColorScale(midValueColor = NULL, maxValueColor = NULL,
  minValueColor = NULL, noDataColor = NULL)
}
\arguments{
\item{midValueColor}{The background color for cells with a color value at the midpoint between}

\item{maxValueColor}{The background color for cells with a color value greater than or equal}

\item{minValueColor}{The background color for cells with a color value less than or equal to}

\item{noDataColor}{The background color for cells that have no color data associated with}
}
\value{
TreemapChartColorScale object
}
\description{
TreemapChartColorScale Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A color scale for a treemap chart.
}
\concept{TreemapChartColorScale functions}
