% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{TreemapChartSpec}
\alias{TreemapChartSpec}
\title{TreemapChartSpec Object}
\usage{
TreemapChartSpec(sizeData = NULL, textFormat = NULL, parentLabels = NULL,
  headerColor = NULL, labels = NULL, colorData = NULL, maxValue = NULL,
  colorScale = NULL, hideTooltips = NULL, hintedLevels = NULL,
  minValue = NULL, levels = NULL)
}
\arguments{
\item{sizeData}{The data that determines the size of each treemap data cell}

\item{textFormat}{The text format for all labels on the chart}

\item{parentLabels}{The data the contains the treemap cells' parent labels}

\item{headerColor}{The background color for header cells}

\item{labels}{The data that contains the treemap cell labels}

\item{colorData}{The data that determines the background color of each treemap data cell}

\item{maxValue}{The maximum possible data value}

\item{colorScale}{The color scale for data cells in the treemap chart}

\item{hideTooltips}{True to hide tooltips}

\item{hintedLevels}{The number of additional data levels beyond the labeled levels to be shown}

\item{minValue}{The minimum possible data value}

\item{levels}{The number of data levels to show on the treemap chart}
}
\value{
TreemapChartSpec object
}
\description{
TreemapChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A <a href='/chart/interactive/docs/gallery/treemap'>Treemap chart</a>.
}
