% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateConditionalFormatRuleRequest}
\alias{UpdateConditionalFormatRuleRequest}
\title{UpdateConditionalFormatRuleRequest Object}
\usage{
UpdateConditionalFormatRuleRequest(index = NULL, sheetId = NULL,
  newIndex = NULL, rule = NULL)
}
\arguments{
\item{index}{The zero-based index of the rule that should be replaced or moved}

\item{sheetId}{The sheet of the rule to move}

\item{newIndex}{The zero-based new index the rule should end up at}

\item{rule}{The rule that should replace the rule at the given index}
}
\value{
UpdateConditionalFormatRuleRequest object
}
\description{
UpdateConditionalFormatRuleRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates a conditional format rule at the given index,or moves a conditional format rule to another index.
}
\concept{UpdateConditionalFormatRuleRequest functions}
