% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateDeveloperMetadataRequest}
\alias{UpdateDeveloperMetadataRequest}
\title{UpdateDeveloperMetadataRequest Object}
\usage{
UpdateDeveloperMetadataRequest(dataFilters = NULL, fields = NULL,
  developerMetadata = NULL)
}
\arguments{
\item{dataFilters}{The filters matching the developer metadata entries to update}

\item{fields}{The fields that should be updated}

\item{developerMetadata}{The value that all metadata matched by the data filters will be updated to}
}
\value{
UpdateDeveloperMetadataRequest object
}
\description{
UpdateDeveloperMetadataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request to update properties of developer metadata.Updates the properties of the developer metadata selected by the filters tothe values provided in the DeveloperMetadata resource.  Callers mustspecify the properties they wish to update in the fields parameter, as wellas specify at least one DataFilter matching the metadata they wish toupdate.
}
