% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateDeveloperMetadataResponse}
\alias{UpdateDeveloperMetadataResponse}
\title{UpdateDeveloperMetadataResponse Object}
\usage{
UpdateDeveloperMetadataResponse(developerMetadata = NULL)
}
\arguments{
\item{developerMetadata}{The updated developer metadata}
}
\value{
UpdateDeveloperMetadataResponse object
}
\description{
UpdateDeveloperMetadataResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response from updating developer metadata.
}
