% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateValuesByDataFilterResponse}
\alias{UpdateValuesByDataFilterResponse}
\title{UpdateValuesByDataFilterResponse Object}
\usage{
UpdateValuesByDataFilterResponse(updatedData = NULL, updatedRows = NULL,
  updatedColumns = NULL, updatedRange = NULL, updatedCells = NULL,
  dataFilter = NULL)
}
\arguments{
\item{updatedData}{The values of the cells in the range matched by the dataFilter after all}

\item{updatedRows}{The number of rows where at least one cell in the row was updated}

\item{updatedColumns}{The number of columns where at least one cell in the column was updated}

\item{updatedRange}{The range (in A1 notation) that updates were applied to}

\item{updatedCells}{The number of cells updated}

\item{dataFilter}{The data filter that selected the range that was updated}
}
\value{
UpdateValuesByDataFilterResponse object
}
\description{
UpdateValuesByDataFilterResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response when updating a range of values by a data filter in aspreadsheet.
}
