% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ValueRange}
\alias{ValueRange}
\title{ValueRange Object}
\usage{
ValueRange(range = NULL, majorDimension = NULL, values = NULL)
}
\arguments{
\item{range}{The range the values cover, in A1 notation}

\item{majorDimension}{The major dimension of the values}

\item{values}{The data that was read or to be written}
}
\value{
ValueRange object
}
\description{
ValueRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data within a range of the spreadsheet.
}
\seealso{
Other ValueRange functions: \code{\link{spreadsheets.values.append}},
  \code{\link{spreadsheets.values.update}}
}
