% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{WaterfallChartColumnStyle}
\alias{WaterfallChartColumnStyle}
\title{WaterfallChartColumnStyle Object}
\usage{
WaterfallChartColumnStyle(label = NULL, color = NULL)
}
\arguments{
\item{label}{The label of the column's legend}

\item{color}{The color of the column}
}
\value{
WaterfallChartColumnStyle object
}
\description{
WaterfallChartColumnStyle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Styles for a waterfall chart column.
}
