% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{WaterfallChartCustomSubtotal}
\alias{WaterfallChartCustomSubtotal}
\title{WaterfallChartCustomSubtotal Object}
\usage{
WaterfallChartCustomSubtotal(subtotalIndex = NULL, dataIsSubtotal = NULL,
  label = NULL)
}
\arguments{
\item{subtotalIndex}{The 0-based index of a data point within the series}

\item{dataIsSubtotal}{True if the data point at subtotal_index is the subtotal}

\item{label}{A label for the subtotal column}
}
\value{
WaterfallChartCustomSubtotal object
}
\description{
WaterfallChartCustomSubtotal Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A custom subtotal column for a waterfall chart series.
}
