% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{WaterfallChartDomain}
\alias{WaterfallChartDomain}
\title{WaterfallChartDomain Object}
\usage{
WaterfallChartDomain(data = NULL, reversed = NULL)
}
\arguments{
\item{data}{The data of the WaterfallChartDomain}

\item{reversed}{True to reverse the order of the domain values (horizontal axis)}
}
\value{
WaterfallChartDomain object
}
\description{
WaterfallChartDomain Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The domain of a waterfall chart.
}
