% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{WaterfallChartSeries}
\alias{WaterfallChartSeries}
\title{WaterfallChartSeries Object}
\usage{
WaterfallChartSeries(subtotalColumnsStyle = NULL,
  positiveColumnsStyle = NULL, data = NULL,
  negativeColumnsStyle = NULL, hideTrailingSubtotal = NULL,
  customSubtotals = NULL)
}
\arguments{
\item{subtotalColumnsStyle}{Styles for all subtotal columns in this series}

\item{positiveColumnsStyle}{Styles for all columns in this series with positive values}

\item{data}{The data being visualized in this series}

\item{negativeColumnsStyle}{Styles for all columns in this series with negative values}

\item{hideTrailingSubtotal}{True to hide the subtotal column from the end of the series}

\item{customSubtotals}{Custom subtotal columns appearing in this series}
}
\value{
WaterfallChartSeries object
}
\description{
WaterfallChartSeries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single series of data for a waterfall chart.
}
\concept{WaterfallChartSeries functions}
