% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{WaterfallChartSpec}
\alias{WaterfallChartSpec}
\title{WaterfallChartSpec Object}
\usage{
WaterfallChartSpec(series = NULL, connectorLineStyle = NULL,
  domain = NULL, firstValueIsTotal = NULL, stackedType = NULL,
  hideConnectorLines = NULL)
}
\arguments{
\item{series}{The data this waterfall chart is visualizing}

\item{connectorLineStyle}{The line style for the connector lines}

\item{domain}{The domain data (horizontal axis) for the waterfall chart}

\item{firstValueIsTotal}{True to interpret the first value as a total}

\item{stackedType}{The stacked type}

\item{hideConnectorLines}{True to hide connector lines between columns}
}
\value{
WaterfallChartSpec object
}
\description{
WaterfallChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A waterfall chart.
}
