% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_to_tab}
\alias{write_to_tab}
\title{Write to Tab}
\usage{
write_to_tab(dta, ss_ids, ranges = NULL, tabs = NULL, cnames = T,
  clear_all = F)
}
\arguments{
\item{dta}{list data that can be coerced to a character matrix}

\item{ss_ids}{vector of sheet ids. If a vector of length 1, will be recycled}

\item{ranges}{if ranges is specified,}

\item{tabs}{character vector of tab names, if specified then the ranges will
be assumed to be size of dataframe}

\item{cnames}{boolean, should colnames be written as well?}

\item{clear_all}{boolean, should the entire sheet be cleared or just the
columns specified in the range. Vectorized or recycled}
}
\value{
writes to gsheet, returns responses
}
\description{
High level wrapper for writing dataframes to tabs of google spreadsheets
}
