% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AppendDimensionRequest}
\alias{AppendDimensionRequest}
\title{AppendDimensionRequest Object}
\usage{
AppendDimensionRequest(dimension = NULL, length = NULL, sheetId = NULL)
}
\arguments{
\item{dimension}{Whether rows or columns should be appended}

\item{length}{The number of rows or columns to append}

\item{sheetId}{The sheet to append rows or columns to}
}
\value{
AppendDimensionRequest object
}
\description{
AppendDimensionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Appends rows or columns to the end of a sheet.
}
