% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BatchGetValuesByDataFilterResponse}
\alias{BatchGetValuesByDataFilterResponse}
\title{BatchGetValuesByDataFilterResponse Object}
\usage{
BatchGetValuesByDataFilterResponse(spreadsheetId = NULL, valueRanges = NULL)
}
\arguments{
\item{spreadsheetId}{The ID of the spreadsheet the data was retrieved from}

\item{valueRanges}{The requested values with the list of data filters that matched them}
}
\value{
BatchGetValuesByDataFilterResponse object
}
\description{
BatchGetValuesByDataFilterResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response when retrieving more than one range of values in a spreadsheetselected by DataFilters.
}
