% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BubbleChartSpec}
\alias{BubbleChartSpec}
\title{BubbleChartSpec Object}
\usage{
BubbleChartSpec(legendPosition = NULL, bubbleOpacity = NULL,
  bubbleSizes = NULL, domain = NULL, bubbleBorderColor = NULL,
  bubbleTextStyle = NULL, groupIds = NULL, bubbleLabels = NULL,
  bubbleMinRadiusSize = NULL, bubbleMaxRadiusSize = NULL, series = NULL)
}
\arguments{
\item{legendPosition}{Where the legend of the chart should be drawn}

\item{bubbleOpacity}{The opacity of the bubbles between 0 and 1}

\item{bubbleSizes}{The data contianing the bubble sizes}

\item{domain}{The data containing the bubble x-values}

\item{bubbleBorderColor}{The bubble border color}

\item{bubbleTextStyle}{The format of the text inside the bubbles}

\item{groupIds}{The data containing the bubble group IDs}

\item{bubbleLabels}{The data containing the bubble labels}

\item{bubbleMinRadiusSize}{The minimum radius size of the bubbles, in pixels}

\item{bubbleMaxRadiusSize}{The max radius size of the bubbles, in pixels}

\item{series}{The data contianing the bubble y-values}
}
\value{
BubbleChartSpec object
}
\description{
BubbleChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A <a href='/chart/interactive/docs/gallery/bubblechart'>bubble chart</a>.
}
