% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{spreadsheets_get}
\alias{spreadsheets_get}
\title{Gets values in a range of a spreadsheet.The caller must specify the spreadsheet ID, range.}
\usage{
spreadsheets_get(data_type = "data.frame", spreadsheetId, range)
}
\arguments{
\item{data_type}{The expected data type of the output}

\item{spreadsheetId}{The ID of the spreadsheet to get}

\item{range}{The A1 notation of the values to get}
}
\description{
Gets values in a range of a spreadsheet.The caller must specify the spreadsheet ID, range.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}
}
