% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{TextRotation}
\alias{TextRotation}
\title{TextRotation Object}
\usage{
TextRotation(vertical = NULL, angle = NULL)
}
\arguments{
\item{vertical}{If true, text reads top to bottom, but the orientation of individual}

\item{angle}{The angle between the standard orientation and the desired orientation}
}
\value{
TextRotation object
}
\description{
TextRotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The rotation applied to text in a cell.
}
