% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateDimensionGroupRequest}
\alias{UpdateDimensionGroupRequest}
\title{UpdateDimensionGroupRequest Object}
\usage{
UpdateDimensionGroupRequest(dimensionGroup = NULL, fields = NULL)
}
\arguments{
\item{dimensionGroup}{The group whose state should be updated}

\item{fields}{The fields that should be updated}
}
\value{
UpdateDimensionGroupRequest object
}
\description{
UpdateDimensionGroupRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the state of the specified group.
}
